// cosmowcl/src/app/shared/models/api/cosmobusiness/istanzaParametroFormLogico.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * cosmobusiness-formlogici.v1
 * API per la gestione dei form logici
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Parametro di una funzionalita
 */
export interface IstanzaParametroFormLogico { 
    chiave?: string;
    descrizione?: string;
    valore?: string;
}